Separated out cs_calibration data load and process tools
