+++ pyMAP/loSim/cs.py
@@ -2,161 +2,74 @@ import pandas as pd
 import numpy as np
 import os
 import periodictable as perd
-
-def load_cal_data(dloc = 'auto',index_cols = 8,data_cols = 27):
-    # Load and perform some preprocessing on elena cal data. Automatically uses package cal results
-    # from pyMAP/cal/cal_results/cs_Ilena_scattering_results.xlsx if no file provided
-    # Input:
-    #   dloc (str): location of cal data file
-    #   index_cols (int): range of columns to use for indexing (not the energy dependent dist values)
-    #   data_cols (int): total range of columns to load containing data
-    # Output: 
-    #   cal (pandas.Dataframe): Illena calibration distribution results
-
-    # locate the scattering data
-    if dloc == 'auto':
-        lpath = os.path.dirname(__file__)
-        for up in range(2):
-            lpath = os.path.dirname(lpath+'..')
-
-        f_cs_data = os.path.join(lpath,"cal/cal_results/CS_Ilena_Scattering_Results.xlsx")
-    else: f_cs_data = dloc
-
-    #import the scattering data
-    hframe = pd.read_excel(f_cs_data,usecols = np.arange(index_cols,data_cols),
-                                    nrows = 5)[3:].T
-    # define header and reduce
-    hframe.values[:,1] = hframe.values[:,1]/2
-    head = pd.MultiIndex.from_frame(hframe,names = ['fit','energy'])
-
-    cal = pd.read_excel(f_cs_data,usecols = np.arange(data_cols),header = 5)
-    nams = ['geo','sample','coating','roughness','location','incident_angle','species','recoil_props']
-    cal[nams] = cal[nams].fillna(method = 'ffill')
-    cal.set_index(nams,inplace = True)
-    cal.set_axis(head,axis = 1,inplace = True)
-    for group in ['automatic','contour']:
-        for val in cal[group].columns:
-            cal[(group,val)] = pd.to_numeric(cal[(group,val)],errors='coerce')
-    
-    # average the cal data for two FWHM finding routines (automatic and contour)
-    #   and over all tested CS locations
-    cal = cal.stack().mean(axis = 1).unstack('location').mean(axis =1).unstack('energy')
-    # cal = cal.stack('energy').unstack('location').median(axis =1).unstack()
-    # drop unneded columns
-    cal = cal.reset_index(['roughness','coating'],drop = True)
-    return(cal)
-
-def setup_cal_data_ke(cal,
-                    use_species=['H','O'], 
-                         ibex_samples=['xxx','L109'], 
-                         imap_samples = ['036b',39] ):
-
-    # Unused Codes
-    #Perform some pre-processing to execute a linear energy-based extrapolation to get
-    #to get 15 deg scattering distribution functions
-    #   cal: cs scattering data to setup
-    #   use_species: the species we will be analyzing
-    #   ibex_samples: Ibex sample effic and scattering distributions to use
-    #   imap_samples: Ibex sample effic and scattering distributions to use
-    #calculate angular scale factors to convert 8 deg to 15 deg
-    # function made laregly obselete by vperp_av_data
-    angs = (cal.stack().unstack(level = 'incident_angle')[15]/cal.stack().unstack(level = 'incident_angle')[8]).unstack()
-    angs_15 = angs.groupby(['recoil_props','species']).mean().drop('effic')
-    angs_convert = angs_15.mean(axis = 1).unstack(level = 'recoil_props')
-    angs_convert['effic'] = 1
-    angs_convert = angs_convert.unstack()
-
-    # drop the 15 deg incident angle observations
-    cal_sml = cal.drop(15,level = 'incident_angle').reset_index(level = 'incident_angle',drop= True)  
-    # Pick out the desired samples to use in simulations
-    use_samples = cal.loc[np.logical_or.reduce([\
-                        np.logical_and(\
-                               cal.reset_index()['geo']==g,
-                               cal.reset_index()['sample']==v)\
-                                          for g,samp in zip(['ibex','imap'],
-                                                    [ibex_samples,imap_samples])\
-                                          for v in samp])]
-    # average scattering data for similar geometries
-    use_dat = use_samples.groupby(['geo','species','recoil_props']).mean()
-    # Scale scattering data up to 15 deg
-    use_dat = use_dat.groupby(['geo','species','recoil_props']).apply(\
-                lambda x: angs_convert.loc[x.reset_index()['recoil_props'],
-                                           x.reset_index()['species']].values[0]*x)
-    #pick out species to use in simulation
-    return(use_dat.stack().unstack('species')[use_species].stack().unstack('energy'))
-
-def setup_fit_funcs_ke(use_dat,cent_eng,e_loss):
-    # Unused Codes
-    def thing(xx):
-        x = xx.dropna()
-        fde = bp.Jonda(xy_data = np.stack([x.keys(),x.values]),func = 'linear')
-        fde.fit_xy(use_err = False)
-        return(fde)
-    #linear fit all the scattering data
-    fits = use_dat.apply(thing,axis =1 )
-
-    #Add energy loss, assuming here that the energy loss for all species and surfaces are the same
-    fde = bp.Jonda(xy_data = np.stack([cent_eng,e_loss]),func = 'linear')
-    fde.fit_xy(use_err = False)
-    fits = fits.unstack(level = 'recoil_props')
-    fits['e_loss'] = [fde]*len(fits)
-    fits = fits.stack()
-    return(fits)
-
-def get_vperp(m_amu,ke,ang):
-    from pyMAP.pyMAP.tof import v_00
-    # Calculate the perpendicular velocity component from the 
-    v = v_00(m_amu,ke)/10**-9/10**6
-    return(np.sin(ang*np.pi/180)*v)
-
-
-def vperp_av_data(cal,samples= ['036b',39,'100P','40P','xxx','L109'],
-                                   mass_split = 10,v_bins = np.linspace(0,90,45)):
-    # use cal data and desired samples to express scattering fwhm distributions using v_perp
-    # rather than ke. scattering distributions should be linear in this form
-
-    # select sample data to be used in determination of cs scattering effects
-    sml_cal = cal.stack().unstack(['sample','geo'])[samples].mean(axis = 1)
-    sml_cal.name = 'val'
-    sml_cal = sml_cal.reset_index()
-    # calculate the v_perp value from mass, energy and incident angle
-    sml_cal['v_perp'] = sml_cal.apply(\
-                    lambda x: get_vperp(perd.elements.symbol(x['species']).mass,
-                                                x['energy'],x['incident_angle']),axis = 1)
-    
-    # define mass groups and v_groups to allow for selective averaging of the sample data
-    sml_cal['m'] = sml_cal['species'].apply(lambda x: perd.elements.symbol(x).mass)
-    sml_cal['m_group'] = ((sml_cal['m'].values/mass_split).astype(int)>=1).astype(int)
-    sml_cal['v_group'] = np.digitize(sml_cal['v_perp'],v_bins)
-    
-    # average sample data and transform to a shape to allow for simple linear fitting
-    sml_cal.set_index(['recoil_props','v_perp','species','v_group'],inplace = True)
-    sml_cal = sml_cal['val']
-    sml_cal = sml_cal.reset_index('v_perp').groupby(['v_group','species','recoil_props']).mean()
-    return(sml_cal.reset_index('v_group',drop = True).set_index('v_perp',append = True).sort_index().reset_index('v_perp'))
-
-def get_cal_fits(load_data_input = {},data_av_input = {}):
-    # function to load elena scattering data, linearize the desired data in v_perp
-    # and perform linear fit functions
-    from pyMAP import bowPy as bp
-
-    vdat = vperp_av_data(load_cal_data(**load_data_input),**data_av_input)
-    def thing(xx):
-        x = xx['v_perp']
-        y = xx['val']
-        fde = bp.Jonda(xy_data = np.stack([x,y]),func = 'linear')
-        fde.fit_xy(use_err = False)
-        fde.name = xx.name
-        return(fde)    
-    fits = vdat.groupby(['species','recoil_props']).apply(thing).unstack('recoil_props')
-    return(fits)
+from .cs_cal import *
 
 def rel_angle(theta1,phi1,theta2,phi2):
     t1 = theta1/90*np.pi/2
     t2 = theta2/90*np.pi/2
     p1 = phi1/90*np.pi/2
     p2 = phi2/90*np.pi/2
-    return(90/np.pi*2*np.arccos(np.sin(t1)*np.sin(t2)+np.cos(t1)*np.cos(t2)*np.cos(p1-p2)))
+    return(np.abs(180/np.pi*np.arccos(np.sin(t1)*np.sin(t2)+np.cos(t1)*np.cos(t2)*np.cos(p1-p2))))
+
+def surf_binding_sputter(E0,cosr,E_bind):
+    from simPyon.simPyon.particles import pdf
+    def dn_de(E,e_bind,cosr):
+        return(E/(E+e_bind)**3*cosr)
+
+    e_pdf = pdf()
+    e_pdf.f = dn_de
+    e_pdf.kwargs = {'e_bind':E_bind,
+                    'cosr':cosr}
+    return(e_pdf.sample(len(E0),0,np.max(E0)))
+
+def recoil_sputter_inellastic(E0,E1,theta1,phi1,theta2,phi2,m1,m2,
+                                    a = 1,b = 1,c = 1,mean_E_loss = .15):
+    '''
+    calculation of the recoil sputtered particle energy distribution 
+    - E0: incidention Neutral energy
+    - E1: recoil ion energy (determined thorugh the usual way assigning a randomized energy loss process)
+    - theta1: incident neutral elevation angle
+    - phi1: incident neutral azimuth angle
+    - theta1: recoil ion elevation angle
+    - phi1: recoil ion azimuth angle
+    - M1: mass of incident neutral
+    - M2: mass of sputtered ion
+    - mean_E_loss: average energy loss experessed as a fraction of the incident neutral energy
+    '''
+    mu = m2/m1
+    gamma = 4*mu/(1+mu)**2
+    rel_ang = rel_angle(theta1,phi1,theta2,phi2)
+    cosr = np.cos(np.abs(rel_ang*np.pi/180))
+    from simPyon.simPyon.particles import source
+    e_loss = source('poisson',{'a1':0,
+                                        'b1':4,
+                                        'c':0,
+                                        'b':.4,
+                                        'k':1,
+                                        'fwhm':1,
+                                        'mean':1,
+                                        'direction':1})
+    q = e_loss(len(E0))*cosr/np.max(cosr)*E0*mean_E_loss
+    c1 = 2*mu/(1+mu)**2*E0*cosr**2
+    c2 = 2*cosr/(1+mu)*np.sqrt(np.abs((mu/(1+mu)*E0*cosr)**2-mu*E0*q/(1+mu)))
+    c3 = -q/(1+mu)
+    Er = (a*c1+b*c2+c*c3)
+    return(np.abs(Er))
+
+
+def Er_sputter(E0,E1,theta1,phi1,theta2,phi2,m1,m2,E_bind=6.9):
+    rel_ang = rel_angle(theta1,phi1,theta2,phi2)
+    cosr = np.cos(np.abs(rel_ang*np.pi/180))
+    # calculaate the recoil scattering energy for the knock on sputtered component
+    Er = recoil_sputter_inellastic(E0,E1,theta1,phi1,theta2,phi2,m1,m2)
+    Er[rel_ang>90]=surf_binding_sputter(E0[rel_ang>90],cosr[rel_ang>90],E_bind)
+    return(Er)
+
+def Er_ellastic(E0,E1,theta1,phi1,theta2,phi2):
+    mu = m2/m1
+    gamma = 4*mu/(1+mu)**2
+    cosr = np.cos(rel_angle(theta1,phi1,theta2,phi2)*np.pi/180)
+    return(gamma*E0*cosr**2)
 
 class cs_scatterer:
     # Control structure for cs scattering montecarlo simulations
@@ -176,22 +89,25 @@ class cs_scatterer:
         self.type = 'modulator'
         self.sputtering = frac_sputtered
         self.is_sputtered = None
+        self.sputtered_m= 1
 
         # assign distribution functions and cs scattering modulator functions
         self.ke = {
-                   'pdf': sim.particles.pdf('poisson',{'c':0,'b':.155,'k':1}),
-                    'pdf2':sim.particles.pdf('sputtered'),
+                   'pdf': sim.particles.pdf('poisson',{'c':0,'b':.105,'k':1}),
+                    'pdf_sputter':sim.particles.pdf('sputtered'),
                    'modulator_f': self.cal_fits['e_loss'][self.species],
                     }
         self.theta = {
                        'pdf': sim.particles.pdf('poisson',{'c':0,'b':.405,'k':1}),
-                       'pdf2':sim.particles.source('cos',
-                                    dist_vals = {'mean':cs_elevation -90,'range':180,'a':0,'b':180,'x_min':0}),
+                       'pdf_sputter':sim.particles.source('cos',
+                                    dist_vals = {'mean':cs_elevation-90,'range':180,'a':90,'b':180,'x_min':0}),
                        'modulator_f': self.cal_fits['theta'][self.species],
                     }
         self.phi = {
                        'pdf': sim.particles.source('gaussian',{'mean':0,'fwhm':1}),
-                       'pdf2':sim.particles.source('cos',
+                       # 'pdf_sputter':sim.particles.source('uniform',
+                       #              dist_vals = {'min':-90,'max':90}),
+                       'pdf_sputter':sim.particles.source('cos',
                                     dist_vals = {'mean':0,'range':180,'a':0,'b':180,'x_min':0}),
                        'modulator_f': self.cal_fits['phi'][self.species],
                     }
@@ -211,18 +127,12 @@ class cs_scatterer:
         direction = -1
         
         new_ke = (self.ke['pdf'].sample(len(ke),0,4)-self.ke['pdf']['b'])*fwhm*direction+mean
-        # take the values that show up below 0 and make them sputtered
-        # try:
-        #     new_ke[new_ke<0] = self.ke['pdf2'].sample(np.sum(new_ke<0),0,.67)*ke[new_ke<0]
-        # except:
-        #     new_ke = abs(new_ke)
-
-        # take the stat determined sputtered stuff and sample the sputtered distribution
-        # new_ke = abs(new_ke)
+        # take the values that show up below 0 and mark them as sputtered
         neg_log = new_ke<0
         self.is_sputtered[neg_log] = True
-        if any(self.is_sputtered):
-            new_ke[self.is_sputtered] = self.ke['pdf2'].sample(np.sum(self.is_sputtered),0,.67)*ke[self.is_sputtered]
+        # if any(neg_log):
+        #     new_ke[neg_log] = self.ke['pdf2'].sample(np.sum(neg_log),0,.67)*ke[neg_log]
+            # new_ke[self.is_sputtered] = self.ke['pdf2'].sample(np.sum(self.is_sputtered),0,.67)*ke[self.is_sputtered]
         return(new_ke)
     
     def phi_scatter(self,ke,theta,phi):
@@ -239,7 +149,7 @@ class cs_scatterer:
 
         # take the stat determined sputtered stuff and sample the sputtered distribution
         if any(self.is_sputtered):
-            phi_new[self.is_sputtered] = self.phi['pdf2'](np.sum(self.is_sputtered))
+            phi_new[self.is_sputtered] = self.phi['pdf_sputter'](np.sum(self.is_sputtered))
         return(phi_new)
 
     def theta_scatter(self,ke,theta,phi):
@@ -255,7 +165,7 @@ class cs_scatterer:
         theta_new = (self.theta['pdf'].sample(len(theta),0,4)-self.theta['pdf']['b'])*fwhm*direction+mean
         # take the stat determined sputtered stuff and sample the sputtered distribution
         if any(self.is_sputtered):
-            theta_new[self.is_sputtered] = self.theta['pdf2'](np.sum(self.is_sputtered))
+            theta_new[self.is_sputtered] = self.theta['pdf_sputter'](np.sum(self.is_sputtered))
         return(theta_new)
 
     def conv_effic(self,ke,theta,phi):
@@ -291,24 +201,31 @@ class cs_scatterer:
         self.is_sputtered = np.random.rand(len(source_df))<self.sputtering
 
         if type(data) != sim_data:
-            data['ke'] = self.ke_scatter(ke,theta,phi)
             data['theta'] = self.theta_scatter(ke,theta,phi)
             data['phi'] = self.phi_scatter(ke,theta,phi)
+            data['ke'] = self.ke_scatter(ke,theta,phi)
             data['counts'] = self.conv_effic(ke,theta,phi)/100
             self.data = data[good_cols]
-            return(self.data)
+            
         else: 
             splat = data.df.copy()[good_cols]
-            splat['ke'] = self.ke_scatter(ke,theta,phi)
             splat['theta'] = self.theta_scatter(ke,theta,phi)
             splat['phi'] = self.phi_scatter(ke,theta,phi)
+            splat['ke'] = self.ke_scatter(ke,theta,phi)
             splat['counts'] = self.conv_effic(ke,theta,phi)/100
             splat['is_start'] = False
+            # take the sputtered particles and input the coupled sputtering energy-angle distribution
+            splat['ke'][self.is_sputtered] = Er_sputter(ke,splat['ke'],
+                                        theta,phi,
+                                        splat['theta'],splat['phi'],
+                                        self.m,self.sputtered_m)[self.is_sputtered]
             data.df = pd.concat([data.df[good_cols],splat],
                             axis = 0).set_index('ion n',
                             append = True).sort_index().reset_index(level = 'ion n')
             self.data = data
-            return(self.data)
+
+
+        return(self.data)
 
     def fly_trajectory(self,source_df,fig,ax):
         # dat = self.fly(source_df)
